update : Code changes based on updated binding files
